fix cuda compile error with nse_net (#1178)
this refactors hybrj to no longer take the function and Jacobian as argument
