+++ nse_solver/nse_index.cpp
@@ -2,7 +2,7 @@
 
 // Declare extern nse index variables
 
-bool NSE_INDEX::initialized;
+AMREX_GPU_MANAGED bool NSE_INDEX::initialized;
 AMREX_GPU_MANAGED int NSE_INDEX::p_index {-1};
 AMREX_GPU_MANAGED int NSE_INDEX::h1_index {-1};
 AMREX_GPU_MANAGED int NSE_INDEX::n_index {-1};
