+++ util/hybrj/hybrj.H
@@ -12,19 +12,17 @@
 #include <AMReX_GpuUtility.H>
 
 template<int neqs, typename T>
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
 void hybrj(hybrj_t<neqs>& hj,
-           const T& data,
-           std::function<void(amrex::Array1D<Real, 1, neqs>&,
-                              amrex::Array1D<Real, 1, neqs>&,
-                              const T&, int&)> fcn,
-           std::function<void(amrex::Array1D<Real, 1, neqs>&,
-                              amrex::Array2D<Real, 1, neqs, 1, neqs>&,
-                              const T&, int&)> jcn) {
+           const T& data) {
 
     // the purpose of hybrj is to find a zero of a system of
     // n nonlinear functions in n variables by a modification
     // of the powell hybrid method. the user must provide a
     // subroutine which calculates the functions and the jacobian.
+    //
+    // We require fcn and jcn to be defined in the other header file
+    // where we include this header.
     //
     // fcn jcn are the names of the user-supplied subroutine which
     // calculates the functions and the jacobian.
